AceComm-3.0: - Bump to v7 - Add autodetection of 4.1, where we make use of RegisterAddonMessagePrefix(). - Pull in ChatThrottleLib v22 (4.1 aware) tests/AceComm: - Make new AceComm test cases for 4.1, keep 4.0 ones also! tests/wow_api: - Implement RegisterAddonMessagePrefix
Details: Pre-4.1, we added control bytes to the PREFIX for handling multipart messages. We can't do this in 4.1 anymore since there's a global limit on the number of active prefixes. Now we put the control bytes at the start of multipart messages instead. If a nonmultipart message start with any of these control bytes, they get escaped.